+++ quantgov/corpora/utils.py
@@ -3,7 +3,6 @@ quantgov.corpora.utils - utility functions for the corpus submodule
 """
 import sys
 
-from decorator import decorator
 from pathlib import Path
 
 
@@ -15,17 +14,3 @@ def load_driver(corpus):
     from driver import driver
     sys.path.pop(0)
     return driver
-
-
-@decorator
-def check_nltk(func, *args, **kwargs):
-    if args[-1] is None:
-        raise RuntimeError('Must install NLTK to use {}'.format(func))
-    return func(*args, **kwargs)
-
-
-@decorator
-def check_textblob(func, *args, **kwargs):
-    if args[-2] is None:
-        raise RuntimeError('Must install textblob to use {}'.format(func))
-    return func(*args, **kwargs)
