+++ setup.cfg
@@ -3,7 +3,7 @@ universal = 1
 [tool:pytest]
 addopts = --flake8
 flake8-ignore = 
-    *.py W391 W503 F821
+    *.py W391 W503
     */__init__.py F401
     tests/* F401 E402
